feat(vue): 支持 Vue 语义定位 (tree-sitter + LSP)
- 升级 tree-sitter 0.22→0.26，同步升级所有 grammar crate
  rust 0.21→0.24, python 0.21→0.25, typescript 0.21→0.23, go 0.21→0.25
- 添加 tree-sitter-vue-updated 0.1.0（兼容 tree-sitter >=0.21）
- treesitter/languages.rs: 新增 TsLanguage::Vue 变体
  detect_language: .vue → "vue"
  ts_language: "vue" → Some(Vue)
  grammar_for: Vue → tree_sitter_vue_updated::language()
- 修复 tree-sitter 0.26 breaking changes:
  grammar_for: language() → LANGUAGE.into()（rust/python/go/ts）
  helpers.rs + verify_syntax.rs: node.child(i usize) → node.child(i as u32)
- tests/ts_symbols.rs: 新增 Vue 检测和 SFC 解析无 panic 测试
- cargo check: 0 errors 0 warnings, cargo test: 全绿

+++ seeyue-mcp/src/tools/verify_syntax.rs
@@ -143,7 +143,7 @@ fn collect_errors(node: tree_sitter::Node, src: &[u8], out: &mut Vec<SyntaxIssue
         });
     }
 
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect_errors(child, src, out);
         }
